Split to analyzer and rawtodigi source files
